Refactor S3DB layer
