+++ vtm-android-example/src/org/oscim/android/test/Samples.java
@@ -124,7 +124,7 @@ public class Samples extends Activity {
 
         linearLayout.addView(createLabel("Experiments"));
         linearLayout.addView(createButton(ReverseGeocodeActivity.class));
-        linearLayout.addView(createButton(S3DBMapActivity.class));
+        linearLayout.addView(createButton(OpenScienceMapS3DBMapActivity.class));
         linearLayout.addView(createButton(ThemeStylerActivity.class));
         linearLayout.addView(createButton(JeoIndoorMapActivity.class));
     }
